+++ src/store/actions.js
@@ -1,14 +1,11 @@
 import {offers} from '../mocks/offers';
 
 export const getCoords = (city) => {
-  if (offers.find((item) => item.city === city) === undefined) {
-    return null;
-  }
-  return offers.find((item) => item.city === city).coords;
+  const offer = offers.find((item) => item.city === city);
+
+  return offer === undefined ? null : offers.find((item) => item.city === city).coords;
 };
 
 export const getOffers = (city) => {
   return offers.filter((item) => item.city === city);
 };
-
-export const test = (a) => a + 1;
